Hard Fault helper and fix for freertos runtime list
This helps to determine where did the code crash after the crash, the
code will try to recover the previous values of the registers and the
most important one in this case is the pc register, with that one we
can get the las function in which the code was before the Hard Fault
interrupt.

It also includes a fix for the runtime register failing to start at
the begining of the program.